requirements.txt
streamlit>=1.0.0
numpy
opencv-python>=4.2.0.32
Pillow